requirements.txt
coreapi==2.3.3
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
dj-rest-auth==5.0.1
coverage==7.3.1
markdown==3.4.4
django-filter==23.3
psycopg==3.1.11
zappa==0.58.0
